+++ rest/src/models/models.go
@@ -7,7 +7,7 @@ import (
 )
 
 type User struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        uint64         `gorm:"primaryKey" json:"id"`
 	Name      string         `json:"name"`
 	Email     string         `gorm:"unique;not null;index" json:"email"`
 	Posts     []Post         `json:"posts"`
@@ -20,38 +20,38 @@ type User struct {
 }
 
 type Post struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        uint64         `gorm:"primaryKey" json:"id"`
 	Title     string         `json:"title"`
 	Content   string         `json:"content"`
 	Comments  []Comment      `json:"comments"`
-	UserID    uint           `gorm:"not null; index" json:"userId"`
+	UserID    uint64         `gorm:"not null; index" json:"userId"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `json:"deletedAt"`
 }
 
 type Comment struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        uint64         `gorm:"primaryKey" json:"id"`
 	Content   string         `json:"content"`
-	UserID    uint           `gorm:"not null" json:"userId"`
-	PostID    uint           `gorm:"not null" json:"postId"`
+	UserID    uint64         `gorm:"not null" json:"userId"`
+	PostID    uint64         `gorm:"not null" json:"postId"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `json:"deletedAt"`
 }
 
 type Message struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        uint64         `gorm:"primaryKey" json:"id"`
 	Content   string         `json:"content"`
-	UserID    uint           `gorm:"not null" json:"userId"`
-	ChatID    uint           `gorm:"not null" json:"chatId"`
+	UserID    uint64         `gorm:"not null" json:"userId"`
+	ChatID    uint64         `gorm:"not null" json:"chatId"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `json:"deletedAt"`
 }
 
 type Chat struct {
-	ID           uint           `gorm:"primaryKey" json:"id"`
+	ID           uint64         `gorm:"primaryKey" json:"id"`
 	Messages     []Message      `json:"messages"`
 	Participants []*User        `gorm:"many2many:participants;" json:"participants"`
 	CreatedAt    time.Time      `json:"createdAt"`
